move services & models into dedicated folders
