+++ internal/api/front/perses/src/app/project/dashboard/model/dashboard.model.ts
@@ -11,8 +11,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { ProjectMetadata } from '../../shared/model/api/v1/metadata.model';
-import { HeaderModel } from '../../shared/model/api/v1/kind.model';
+import { ProjectMetadata } from '../../../shared/model/api/v1/metadata.model';
+import { HeaderModel } from '../../../shared/model/api/v1/kind.model';
 
 export enum VariableKind {
   promqlQuery = 'PromQLQuery',
