+++ internal/api/front/perses/src/app/project/dashboard/model/ngxcharts.model.ts
